Replace int with long for return value of solve
adapt main

+++ day_11/java/app/src/main/java/tkrippes/com/github/adventofcode2024/day11/Main.java
@@ -13,7 +13,7 @@ public class Main {
         System.out.println("Result of day 11 part 2: " + result(75));
     }
 
-    private static int result(int numberOfBlinks) {
+    private static long result(int numberOfBlinks) {
         List<PlutonianPebble> stones;
         try {
             stones = PlutonianPebblesParser.parse(inputFileName);
